backend/pyproject.toml
[tool.poetry]
name = "teqsmartsubmit-backend"
version = "0.1.0"
description = "FastAPI backend and automation engine for TEQSmartSubmit"
authors = ["TEQTOP Automation Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
httpx = "^0.27.0"
playwright = "^1.44.0"
pandas = "^2.2.1"
reportlab = "^4.2.0"
redis = "^5.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
ruff = "^0.4.4"
black = "^24.3.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

